# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "get": {
                "tags": [
                    "Authentikasi"
                ],
                "summary": "Get the list of resources",
                "responses": {
                    "200": {
                        "description": "Return a list of resources"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentikasi"
                ],
                "summary": "Logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/bookmark": {
            "get": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Get Bookmark",
                "operationId": "624baa971ef2dce3594bb2801425422d",
                "responses": {
                    "200": {
                        "description": "Bookmark"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "PUT Bookmark",
                "operationId": "c4d724c4c975054d14cd819d963d7aae",
                "responses": {
                    "200": {
                        "description": "PUT Bookmark"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "POST Bookmark",
                "operationId": "66a96659780ea38427e57a1d323992a4",
                "responses": {
                    "200": {
                        "description": "POST Bookmark"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "DELETE Bookmark",
                "operationId": "13f862f1f003d54192bb5049cb77a821",
                "responses": {
                    "200": {
                        "description": "DELETE Bookmark"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/bookmark/{id}": {
            "get": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "GET Bookmark",
                "operationId": "a16a7bdc13e31558badf21e3c490eccf",
                "responses": {
                    "200": {
                        "description": "GET Bookmark"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/kategori": {
            "get": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Get the list of resources",
                "operationId": "2338314090a81813db8257437b649a2f",
                "responses": {
                    "200": {
                        "description": "Return a list of resources"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Tambah Data Kategori",
                "operationId": "updatePetWithForm",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "nama_kategori": {
                                        "description": "Nama Kategori",
                                        "type": "string"
                                    },
                                    "detail": {
                                        "description": "Detail Kategori",
                                        "type": "string"
                                    },
                                    "berkas": {
                                        "description": "Berkas",
                                        "type": "string"
                                    },
                                    "isPembimbing": {
                                        "description": "Pembimbing Boolean",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sukses Ditambahkan"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/kategori/{id}": {
            "get": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Get kategori by kategori id",
                "operationId": "4e5f418e3df1ff50fadd87ced238b8eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Deletes a kategori",
                "operationId": "deletekategori",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "200": {
                        "description": "Data Berhasil Dihapus"
                    },
                    "404": {
                        "description": "kategori not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}